Added MapKit to the storyboard. Added MapKit code to render the annotations. Clicking on any annotation will open the link in safari.

+++ OnTheMap/StudentInfo.swift
@@ -14,15 +14,15 @@ struct StudentInfo {
     var createdAt: String
     var firstName: String
     var lastName: String
-    var latitude: Float
-    var longitude: Float
+    var latitude: Double
+    var longitude: Double
     var mapString: String
     var mediaURL: String
     var objectId: String
     var uniqueKey: String
     var updatedAt: String
     
-    init(createdAt: String, firstName: String, lastName: String, latitude: Float, longitude: Float, mapString: String,
+    init(createdAt: String, firstName: String, lastName: String, latitude: Double, longitude: Double, mapString: String,
          mediaURL: String, objectId: String, uniqueKey: String, updatedAt: String) {
             self.createdAt = createdAt
             self.firstName = firstName
